move common tasks to matrix/common.rs
